refactor(codegen): extrai passes/async_expand.rs de func.rs
Move expand_async_functions + expand_await_exprs (404 linhas) para
crates/rts-codegen/src/codegen/lower/passes/async_expand.rs:
- expand_async_functions: desugar `async function f` → wrapper +
  __async_inner_f via promise.create (#359 followup, design @drysius)
- expand_await_exprs: `await x` → `promise.wait(x)` (#414, F3 #411)

Limpa tambem ~50 linhas de doc-comments orfaos que sobravam apos
extracoes anteriores. func.rs: 6713 -> 6255 linhas. Suite verde.
